face_pipeline/requirements.txt
onnxruntime
numpy
opencv-python
Pillow
scikit-learn
matplotlib